requirements.txt
fastapi
pandas
pdfplumber
pydantic
python-dotenv
uvicorn
